Move SetMinecraftVersion to components (#78)
