+++ packages/cli/src/components/SetMinecraftVersion.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react"
 import SelectInput, { Item } from "ink-select-input"
 import { useState } from "react"
-import { checkForAssets, detectVersions } from "../../../utils"
+import { checkForAssets, detectVersions } from "../utils"
 import { Box, Text } from "ink"
-import { CACHE } from "../../../main"
-import { MinecraftUtility } from "../../../minecraft"
+import { CACHE } from "../main"
+import { MinecraftUtility } from "../minecraft"
 
 const minecraftAssetReader = new MinecraftUtility()
 export interface Item<V> {
@@ -28,37 +28,38 @@ export const SetMinecraftVersion = (props: {
   }
 
   useEffect(() => {
-    /**
-     * @see https://stackoverflow.com/questions/53332321/react-hook-warnings-for-async-function-in-useeffect-useeffect-function-must-ret
-     */
-    async function setData() {
-      let minecraftVersionsArray: any
+    let isCancelled = false
+    let minecraftVersionsArray: any
 
-      detectVersions()
-        .then((v) => (minecraftVersionsArray = v))
-        .then(() => {
+    detectVersions()
+      .then((v) => (minecraftVersionsArray = v))
+      .then(() => {
+        if (!isCancelled) {
           setMinecraftVersions(minecraftVersionsArray)
-        })
-        .then(() => {
-          checkForAssets(selectedVersion).then((path) => {
-            if (path) {
-              try {
-                minecraftAssetReader.readInRawData({
-                  path,
-                })
-                CACHE.setRootAssetsPath(path)
-                props.setRawAssetsPathHandler(path)
-              } catch (e) {
-                console.log(`Unable to read in raw data: `, e.message)
-              }
-            } else {
-              console.error(`path did not exist when passed to readInRawData`)
+        }
+      })
+      .then(() =>
+        checkForAssets(selectedVersion).then((path) => {
+          if (path) {
+            try {
+              minecraftAssetReader.readInRawData({
+                path,
+              })
+              CACHE.setRootAssetsPath(path)
+              props.setRawAssetsPathHandler(path)
+            } catch (e) {
+              console.log(`Unable to read in raw data: `, e.message)
             }
-          })
+          } else {
+            console.error(`path did not exist when passed to readInRawData`)
+          }
         })
-    }
+      )
 
-    setData()
+    // Used to prevent attempting to set state when component is unmounted
+    return () => {
+      isCancelled = true
+    }
   }, [selectedVersion, minecraftVersions])
 
   return (
